pull out inflight checkout stuff to extension
